Fix `Read_constraint()` segfault
The function had a print statement that reported how many disulfides were read and then reported the third one. This results in a segfault if the number of specified disulfides is fewer than 3. I change this to check that there is  more than 1 disulfide constraint and to print the value of the first one.

Some neatening up of the code is also done, with white spacing, braces.