Revise `Compute_constraint_energy` for CB-CB dist
The previous implementation used CA-CA distance in the constraint calculation. Here, this is switched to CB-CB. The potential is implemented as a flat-bottom potential with a flat region between 2.9 and 4.6 A (see Hazes et al. 1988). The potential increases quadratically outside the region but is limited to 0 for distances > 5.6 A. The logic of one residue being involved in only one disulfide is retained.