constraint fix in Gly-handling
My decision for disulfide CB-CB restraints involving Gly residues (in the circumstance that one calls for a CB-CB restraint involving a Gly residue) is that I would just take the CA position of Gly. Unfortunately, here was a bug: for the second residue considered in a user-specified disulfide, I was taking the index of the CA of residue 1, not of residue 2.