Disulfide constraints (restraints) always on
For my purposes, I want a potential that does not go to 0 when the CB-CB distance is beyond some threshold. I instead put in in a linear potential term in addition to the flat well. This means however that the disulfide potential is always active, and so in this commit, I removed the logic of picking the shortest CB-CB distance and having that be a disulfide.