Replica exchange: don't attempt more than once
As previously implemented, for a fixed number of exchange attempts, a
random replica is selected for exchange. Here, I revise this such that
if an exchange with a replica has been attempted before, no further
attempts would be made. This does count toward the fixed number of
exchange attempts, but this prevents multiple tries to exchange
between two pairs.